Add files via upload
To run the simulation execute test.py from the shell or just import InductionMachine from induction_machine.py and execute the function runSim(). 
The code should simulate the dynamics of an induction motor when accelerating to 200 km/h and when climbing on a small hill and then going down at about 30 km/h. For the parameters I tried to recreate the early version of Tesla model S. 
The class is in induction_machine.py; park.py contains the functions for the Park transform and for the inverse Park transform; ramps.py computes the values for the magnitude and frequency reference of the inverter voltage; voltages.py computes the voltage waveforms that feed the motor once for the whole simulation; quickV computes the voltage waveforms that feed the motor at each step of the for loop in auto_model; utilities.py has a collection of functions that interact with the user to choose the feeding voltage waveform to the motor and to plot desired functions; machine_model.py is the model of the motor; auto_model.py is the model of the motor that uses feedback to keep the speed constant; test.py is just a shortcut to execute the simulation quickly from the shell.
I should say that in the class the functions are called many times because auto_model and machine_model each simulate a second of the simulation, so for 6 seconds, they need to be called six times.
I wrote the functions to try learning using classes in python, so they can come off as very inelegant. My aim in uploading these files is to give qualitative insight to the inner workings of an electric motor using an interesting example to people that are curious about electric motors but are not trained electrical engineers. Enjoy! :)